Fix recording rule errors the label values change (#6)
This fixes the following problem:

You define an SLO with label mylabel=original. Let that run for a while. Then you change the label to mylabel=changed. This will cause the slo:sli_error:ratio_rate<window> to fail for the duration of the window with the error vector contains metrics with the same labelset after applying rule labels.

To fix this problem we essentially discard the labels from the query. They are then re-applied to the resulting recording rule.

Note that this uses sub-queries so you need to be running a version of Prometheus that supports that.